# openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:7950","description":"Generated server url"}],"tags":[{"name":"1.회원 가입 API","description":"사용자 회원 가입을 처리하는 API"}],"paths":{"/auth/refresh-token":{"post":{"tags":["login-controller"],"operationId":"refreshToken","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorDTO"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDTOVoid"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDTOAuthResponseDTO"}}}}}}},"/auth/logout":{"post":{"tags":["login-controller"],"operationId":"logout","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorDTO"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDTOVoid"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDTOVoid"}}}}}}},"/auth/login":{"post":{"tags":["login-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequestDTO"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorDTO"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDTOVoid"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDTOAuthResponseDTO"}}}}}}},"/api/register/member":{"post":{"tags":["1.회원 가입 API"],"summary":"일반 회원 가입","description":"이메일, 비밀번호, 이름 등으로 회원 가입을 요청합니다.","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterMemberDTO"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorDTO"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDTOVoid"}}}},"201":{"description":"회원가입 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDTOVoid"}}}}}}},"/api/MyPage/{memberId}/change":{"post":{"tags":["my-page-controller"],"operationId":"changeMemberInformation","parameters":[{"name":"memberId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MemberDTO"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorDTO"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDTOVoid"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDTOVoid"}}}}}}},"/api/MyPage/reservation":{"post":{"tags":["my-page-controller"],"operationId":"reserveCar","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReservationDTO"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorDTO"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDTOVoid"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDTORentDTO"}}}}}}},"/api/rentcars/rank":{"get":{"tags":["rent-cars-controller"],"operationId":"getRankedRentCars","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorDTO"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDTOVoid"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDTOPageRentCarsDTO"}}}}}}},"/api/protected":{"get":{"tags":["protected-api-controller"],"operationId":"protectedEndpoint","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorDTO"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDTOVoid"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/MyPage/{memberId}":{"get":{"tags":["my-page-controller"],"operationId":"memberInformation","parameters":[{"name":"memberId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorDTO"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDTOVoid"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDTOMemberDTO"}}}}}}},"/api/MyPage/reservation/list":{"get":{"tags":["my-page-controller"],"operationId":"getReservationList","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorDTO"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDTOVoid"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDTOListRentDTO"}}}}}}},"/api/MyPage/reservation/list/{reservationId}":{"get":{"tags":["my-page-controller"],"operationId":"getReservation","parameters":[{"name":"reservationId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorDTO"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDTOVoid"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDTORentDTO"}}}}}}},"/api/MyPage/reservation/list/cancel/{reservationId}":{"delete":{"tags":["my-page-controller"],"operationId":"cancelReservation","parameters":[{"name":"reservationId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorDTO"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDTOVoid"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDTOVoid"}}}}}}}},"components":{"schemas":{"ErrorDTO":{"type":"object","properties":{"message":{"type":"string"},"code":{"type":"string"}}},"ApiResponseDTOVoid":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"type":"object"}}},"ApiResponseDTOAuthResponseDTO":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/AuthResponseDTO"}}},"AuthResponseDTO":{"type":"object","properties":{"token":{"type":"string"}}},"LoginRequestDTO":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"RegisterMemberDTO":{"required":["address","email","name","password","phone"],"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"password":{"pattern":"^(?=.*[0-9])(?=.*[!@#$%^&*()_+\\-=\\[\\]{};':\"\\\\|,.<>/?])(?=.*[A-Za-z])[A-Za-z0-9!@#$%^&*()_+\\-=\\[\\]{};':\"\\\\|,.<>/?]{8,16}$","type":"string","writeOnly":true},"phone":{"type":"string"},"address":{"type":"string"}}},"MemberDTO":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"licence":{"type":"boolean"},"email":{"type":"string"},"phone":{"type":"string"},"address":{"type":"string"}}},"RentCarsDTO":{"required":["name","rentPrice"],"type":"object","properties":{"name":{"type":"string"},"rentPrice":{"minimum":0,"type":"integer","format":"int32"},"recommend":{"type":"integer","format":"int64"},"rentCarNumber":{"type":"string"},"reservationStatus":{"type":"string","enum":["AVAILABLE","RENTED","MAINTENANCE"]},"totalPrice":{"minimum":0,"type":"integer","format":"int32"}}},"RentDTO":{"required":["rentCars"],"type":"object","properties":{"rent_id":{"type":"string","format":"uuid"},"rentTime":{"type":"string","format":"date-time"},"duration":{"minimum":1,"type":"integer","format":"int32"},"endTime":{"type":"string","format":"date-time"},"rentCars":{"$ref":"#/components/schemas/RentCarsDTO"}}},"ReservationDTO":{"required":["rentCarsDTO","rentDTO"],"type":"object","properties":{"rentDTO":{"$ref":"#/components/schemas/RentDTO"},"rentCarsDTO":{"$ref":"#/components/schemas/RentCarsDTO"}}},"ApiResponseDTORentDTO":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/RentDTO"}}},"Pageable":{"type":"object","properties":{"page":{"minimum":0,"type":"integer","format":"int32"},"size":{"minimum":1,"type":"integer","format":"int32"},"sort":{"type":"array","items":{"type":"string"}}}},"ApiResponseDTOPageRentCarsDTO":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/PageRentCarsDTO"}}},"PageRentCarsDTO":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"numberOfElements":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/RentCarsDTO"}},"number":{"type":"integer","format":"int32"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"paged":{"type":"boolean"},"unpaged":{"type":"boolean"},"offset":{"type":"integer","format":"int64"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}}}},"SortObject":{"type":"object","properties":{"direction":{"type":"string"},"nullHandling":{"type":"string"},"ascending":{"type":"boolean"},"property":{"type":"string"},"ignoreCase":{"type":"boolean"}}},"ApiResponseDTOMemberDTO":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/MemberDTO"}}},"ApiResponseDTOListRentDTO":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/RentDTO"}}}}}}}